Move tests to right namespace
Move two subscription tests to the right namespace for that, rather than for publishing.
