+++ JustSaying.IntegrationTests/Fluent/Subscribing/WhenAHandlerThrowsAnExceptionWithNoMonitor.cs
@@ -9,7 +9,7 @@ using NSubstitute;
 using Shouldly;
 using Xunit.Abstractions;
 
-namespace JustSaying.IntegrationTests.Fluent.Publishing
+namespace JustSaying.IntegrationTests.Fluent.Subscribing
 {
     public class WhenAHandlerThrowsAnExceptionWithNoMonitor : IntegrationTestBase
     {
